Fix editing of vocab term literals

+++ lib/ar_collection_setter.rb
@@ -4,10 +4,14 @@ AR_COLLECTION_SETTER = ->(input:, binding:, **) do
 
   # Hard-delete records that are being replaced
   # Any further dependent records must be handled with foreign key constraints
-  if collection.respond_to?(:loaded?) && !collection.loaded?
-    collection.delete_all :delete_all
-  else
-    collection.group_by(&:class).each { |klass, items| klass.where(id: items.map(&:id)).delete_all }
+  if collection.is_a?(ActiveRecord::Associations::CollectionProxy)
+    if collection.respond_to?(:loaded?) && !collection.loaded?
+      collection.delete_all :delete_all
+    else
+      collection.group_by(&:class).each do |klass, items|
+        klass.where(id: items.map(&:id)).delete_all
+      end
+    end
   end
 
   # Don't use the collection= method (setter) so we can return meaningful errors
